add a test and avoid patching
