+++ hvplot/plotting/core.py
@@ -1,3 +1,4 @@
+import itertools
 from collections import defaultdict
 
 import param
@@ -1860,6 +1861,47 @@ class hvPlotTabular(hvPlotBase):
         - Plotly: https://plotly.com/python/text-and-annotations/
         """
         return self(x, y, text=text, kind="labels", **kwds)
+
+
+class hvPlotTabularPolars(hvPlotTabular):
+    def _get_converter(self, x=None, y=None, kind=None, **kwds):
+        import polars as pl
+
+        params = dict(self._metadata, **kwds)
+        x = x or params.pop("x", None)
+        y = y or params.pop("y", None)
+        kind = kind or params.pop("kind", None)
+
+        # Find columns which should be converted for LazyDataFrame and DataFrame
+        if isinstance(self._data, (pl.LazyFrame, pl.DataFrame)):
+            if params.get("hover_cols") == "all":
+                columns = list(self._data.columns)
+            else:
+                possible_columns = [
+                    [v] if isinstance(v, str) else v
+                    for v in params.values()
+                    if isinstance(v, (str, list))
+                ]
+                columns = (
+                    set(self._data.columns) & set(itertools.chain(*possible_columns))
+                ) or {self._data.columns[0]}
+                xs = x if is_list_like(x) else (x,)
+                ys = y if is_list_like(y) else (y,)
+                columns |= {*xs, *ys}
+                columns.discard(None)
+
+        if isinstance(self._data, pl.DataFrame):
+            data = self._data.select(columns).to_pandas()
+        elif isinstance(self._data, pl.Series):
+            data = self._data.to_pandas()
+        elif isinstance(self._data, pl.LazyFrame):
+            data = self._data.select(columns).collect().to_pandas()
+        else:
+            raise ValueError(
+                "Only Polars DataFrame, Series, and LazyFrame are supported"
+            )
+
+        return HoloViewsConverter(data, x, y, kind=kind, **params)
 
 
 class hvPlot(hvPlotTabular):
